+++ fjs/djs/transpiler/module.f.ts
@@ -64,7 +64,7 @@ const transpileWithImports
             const pathsCombine = listMap(pathConcat(dir))(parseModuleResult[1][0])
             const pathsArray = toArray(pathsCombine)
             const contextWithStack = { ...context, stack: { first: path, tail: context.stack } }
-            const x0 = foldStep(foldNextModuleOp)(contextWithStack)(pure(pathsArray))
+            const x0 = foldStep(pure(pathsArray), contextWithStack, foldNextModuleOp)
             return step(
                 x0,
                 contextWithImports => {